Fix race in StreamChannelCodecs()
This fixes issue 2 of #294. The root cause is that the lock held when accessing `obj.Streams` is immediately released and doesn't cover accesses to the channels map.

To fix this somewhat elegantly, wrap the block that needs to access to the map in a function and hold the lock for the entire duration of the function.